Move the cli api into the api subproject
