+++ api/maven-api-cli/src/main/mdo/core-extensions.mdo
