+++ api/pom.xml
@@ -42,6 +42,7 @@
     <module>maven-api-metadata</module>
     <module>maven-api-core</module>
     <module>maven-api-spi</module>
+    <module>maven-api-cli</module>
   </modules>
 
   <properties>
