+++ maven-cli/pom.xml
@@ -34,6 +34,10 @@ under the License.
   <dependencies>
     <dependency>
       <groupId>org.apache.maven</groupId>
+      <artifactId>maven-api-cli</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.maven</groupId>
       <artifactId>maven-embedder</artifactId>
     </dependency>
     <dependency>
