+++ maven-cli/src/main/java/org/apache/maven/cling/invoker/mvn/DefaultMavenInvokerRequest.java
@@ -26,10 +26,10 @@ import java.util.Map;
 
 import org.apache.maven.api.annotations.Nonnull;
 import org.apache.maven.api.cli.Logger;
+import org.apache.maven.api.cli.extensions.CoreExtension;
 import org.apache.maven.api.cli.mvn.MavenInvokerRequest;
 import org.apache.maven.api.cli.mvn.MavenOptions;
 import org.apache.maven.api.services.MessageBuilderFactory;
-import org.apache.maven.cli.internal.extension.model.CoreExtension;
 import org.apache.maven.cling.invoker.BaseInvokerRequest;
 
 import static java.util.Objects.requireNonNull;
