+++ maven-cli/src/main/java/org/apache/maven/cling/invoker/mvn/local/LocalParser.java
@@ -45,13 +45,13 @@ import org.apache.maven.api.Constants;
 import org.apache.maven.api.annotations.Nullable;
 import org.apache.maven.api.cli.ParserException;
 import org.apache.maven.api.cli.ParserRequest;
+import org.apache.maven.api.cli.extensions.CoreExtension;
 import org.apache.maven.api.cli.mvn.MavenInvokerRequest;
 import org.apache.maven.api.cli.mvn.MavenOptions;
 import org.apache.maven.api.cli.mvn.MavenParser;
 import org.apache.maven.cli.CLIManager;
 import org.apache.maven.cli.CLIReportingUtils;
 import org.apache.maven.cli.internal.extension.io.CoreExtensionsStaxReader;
-import org.apache.maven.cli.internal.extension.model.CoreExtension;
 import org.apache.maven.cli.props.MavenPropertiesLoader;
 import org.apache.maven.cling.invoker.mvn.CommonsCliMavenOptions;
 import org.apache.maven.cling.invoker.mvn.DefaultMavenInvokerRequest;
